fix crash due to absence of X11InitThread fix #29 this patch removes the need to call X11InitThread in qgis (so a standard qgis will work)
QGIS only calls X11 related functions (draw on screen) in the main
thread, but the layers are rendered offscreen by the rendering threads
(since multithreaded rendering was introduced)

When using mesa (opengl implementation on CPU) the X11 lib is not called
when rendering offscreen, so issuing opengl calls in the rendering
thread works fine.

When using NVIDIA drivers, the opengl calls must reach the graphics
card, so I guess that the XLib is called at some point, and since XLib
is not thread-safe, the concurent calls by qgis main thread and the
rendering thread to XLib functions cause a crash.

One solution is to call X11InitThread when starting the application
(e.g. first line of main()), Qt provides a mechanism for that. But this
as a performance impact since this function sets a global mutex on all
XLib functions.

The other solution is to use opengl calls only in the main thread, this
is what this patch does. The rendering function in meshlayer send a
signals and waits on a condition, the main tread processes the signal
and create an image with opengl calls that end the wait of the rendering
thread.